+++ test/test_dead_myself.py
@@ -9,7 +9,7 @@ servers_list = [33001]
 
 def check_myself(srv, status):
     myself = srv.myself()
-    assert myself['status_name'] == status
+    assert myself['status'] == status
 
 def test_myself(servers, helpers):
     helpers.wait_for(check_myself, [servers[33001], 'dead'], timeout=5)