+++ test/test_broadcast.py
@@ -24,4 +24,5 @@ def check_status(srv, uri, status):
     assert member['status'] == status
 
 def test_join(servers, helpers):
+    servers[33002].broadcast()
     helpers.wait_for(check_status, [servers[33002], hostname+':33001', 'alive'])