+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - API method `set_payload()` now sets only the given key within payload table
+- Hide internal numeric `status` from public API 
 
 ## [2.0.0] - 2018-04-03
 ### Changed