+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- API method `probe_uri()`
+
 ### Changed
 - API method `set_payload()` now sets only the given key within payload table
 - Hide internal numeric `status` from public API